fix generate_dataset vig

+++ R/generate_dataset.R
@@ -60,8 +60,9 @@ generate_dataset <- function(file = tempfile(fileext = ".zarr"),
     .get_basilisk_env(sd_version)
   ) 
   on.exit(basilisk::basiliskStop(proc))
-  basilisk::basiliskRun(proc, function() {
-    unlink(file, recursive = TRUE)
+  basilisk::basiliskRun(proc, function(file) {
+    if(dir.exists(file))
+      unlink(file, recursive = TRUE)
     dummy_sd <- reticulate::import("dummy_spatialdata")
     sd <- reticulate::import("spatialdata")
     temp <- dummy_sd$generate_dataset(
@@ -76,5 +77,5 @@ generate_dataset <- function(file = tempfile(fileext = ".zarr"),
     temp$write(file)
     message("SpatialData object written to '", file, "'")
     return(file)
-  })
+  }, file = file)
 }
